Disable things while a repo is migrated (#716)
When a repository is either migrating or is already migrated we don't want
to allow any changes made to it. This commit blocks all of the routes that could
potentially change the data if a repository is in migrating or migrated state.

+++ lib/travis/api/v3/services/key_pair/delete.rb
@@ -4,7 +4,10 @@ module Travis::API::V3
 
     def run!
       repository = check_login_and_find(:repository)
+
       private_repo_feature!(repository)
+      return repo_migrated if migrated?(repository)
+
       key_pair = find(:key_pair, repository)
       access_control.permissions(repository).delete_key_pair!
       query.delete(repository) and deleted
